day 12 interface mocking complete
