+++ day_11_12_13/test2/userDB/cover.out
@@ -16,25 +16,25 @@ repos/ZopSmartTest/day_11/test2/userDB/users.go:60.33,62.3 1 0
 repos/ZopSmartTest/day_11/test2/userDB/users.go:66.33,68.3 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:74.35,78.3 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:85.16,87.3 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:91.62,96.16 3 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:99.2,99.12 1 0
+repos/ZopSmartTest/day_11/test2/userDB/users.go:91.62,96.16 3 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:99.2,99.12 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:96.16,98.3 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:102.57,109.16 3 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:112.2,115.18 2 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:130.2,130.12 1 0
+repos/ZopSmartTest/day_11/test2/userDB/users.go:102.57,109.16 3 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:112.2,115.18 2 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:130.2,130.12 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:109.16,111.3 1 0
 repos/ZopSmartTest/day_11/test2/userDB/users.go:115.18,123.17 6 0
 repos/ZopSmartTest/day_11/test2/userDB/users.go:128.3,128.57 1 0
 repos/ZopSmartTest/day_11/test2/userDB/users.go:123.17,125.4 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:133.78,144.16 5 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:147.2,147.18 1 0
+repos/ZopSmartTest/day_11/test2/userDB/users.go:133.78,144.16 5 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:147.2,147.18 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:144.16,146.3 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:150.100,153.31 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:158.2,159.24 2 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:172.2,176.16 4 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:180.2,180.12 1 0
+repos/ZopSmartTest/day_11/test2/userDB/users.go:150.100,153.31 1 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:158.2,159.24 2 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:172.2,176.16 4 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:180.2,180.12 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:153.31,155.3 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:159.24,163.26 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:164.15,165.60 1 0
-repos/ZopSmartTest/day_11/test2/userDB/users.go:166.11,167.57 1 0
+repos/ZopSmartTest/day_11/test2/userDB/users.go:159.24,163.26 1 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:164.15,165.60 1 1
+repos/ZopSmartTest/day_11/test2/userDB/users.go:166.11,167.57 1 1
 repos/ZopSmartTest/day_11/test2/userDB/users.go:176.16,179.3 1 0
